+++ config/app/config.php
@@ -8,8 +8,9 @@ return [
     'cache' => env('APP_CACHE', false),
     'item_type_id' => env('ITEM_TYPE_ID'),
     'item_subtype_id' => env('ITEM_SUBTYPE_ID'),
+    'error_email' => env('ERROR_EMAIL'),
     'cookie_user' => env('SESSION_NAME_USER'),
     'cookie_bearer' => env('SESSION_NAME_BEARER'),
-    'version' => '1.05.1',
-    'release_date' => '25th August 2022'
+    'version' => '1.06.0',
+    'release_date' => '26th August 2022'
 ];